requirements.txt
# Google ADK and AI dependencies
google-adk>=0.1.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.60.0

# Environment and configuration
python-dotenv>=1.0.0

# Utilities and helpers
requests>=2.31.0
pydantic>=2.5.0

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Async support (for advanced examples)
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Data processing (for examples)
pandas>=2.1.0
numpy>=1.24.0

# Logging and monitoring
structlog>=23.2.0 